requirements.txt
customtkinter==5.2.2
Pillow==11.2.1
Requests==2.32.3
yt_dlp==2025.3.31
